refactor: extract helper functions / libs from cached_image_resource.go
